Use %if 0%{?pyver} when checking for the python version
If the python-srpm-macros package is not installed in the machine
where we are running rpmbuild -bs, we get an error message like the
following:

bad %if condition

Using 0%{?pyver} instead to pass this. This will not be an issue during
the build, since the python-srpm-macros package is installed as a
dependency of python?-devel.

Change-Id: If0ac6b6df6aba8c4d7295e03d9cced02c4a3f61f